infra/loesoe/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
itsdangerous
python-dotenv
requests
